create helper function for getting day as string
